# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/employees": {
      "get": {
        "operationId": "EmployeesController_findAll",
        "parameters": [
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeEntity"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "List employees",
        "tags": [
          "employees"
        ]
      },
      "post": {
        "operationId": "EmployeesController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the employee conflicts with existing users."
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Create an employee",
        "tags": [
          "employees"
        ]
      }
    },
    "/employees/{id}": {
      "get": {
        "operationId": "EmployeesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Employee not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Get an employee by id",
        "tags": [
          "employees"
        ]
      },
      "patch": {
        "operationId": "EmployeesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the employee conflicts with existing users."
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Employee not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Update an employee",
        "tags": [
          "employees"
        ]
      },
      "delete": {
        "operationId": "EmployeesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Employee not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Delete an employee",
        "tags": [
          "employees"
        ]
      }
    },
...
    "/hr-profiles": {
      "get": {
        "operationId": "HrProfilesController_findAll",
        "parameters": [
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HrProfileEntity"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "List HR profiles",
        "tags": [
          "hr-profiles"
        ]
      },
      "post": {
        "operationId": "HrProfilesController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHrProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "hrProfile": {
                      "id": "hr-1777718400000-a1b2c3d4",
                      "name": "Priya Sharma",
                      "email": "priya.sharma@gmail.com",
                      "username": "priya.sharma@gmail.com",
                      "phoneNumber": "+91 98765 43210",
                      "department": "HR",
                      "designation": "HR Admin",
                      "location": "",
                      "companyId": "CMP-1001",
                      "companyName": "Stack Builders Pvt Ltd",
                      "status": "Active",
                      "createdBySuperUser": true,
                      "password": "secure123",
                      "createdAt": "2026-05-02T12:30:00.000Z",
                      "updatedAt": "2026-05-02T12:30:00.000Z"
                    },
                    "replacedProfiles": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or company reference is invalid."
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Create an HR profile and replace any existing HR mapped to the same company",
        "tags": [
          "hr-profiles"
        ]
      }
    },
    "/hr-profiles/{id}": {
      "get": {
        "operationId": "HrProfilesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrProfileEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "HR profile not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Get an HR profile by id",
        "tags": [
          "hr-profiles"
        ]
      },
      "patch": {
        "operationId": "HrProfilesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHrProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrProfileEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or conflicting company/email details."
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "HR profile not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Update an HR profile",
        "tags": [
          "hr-profiles"
        ]
      },
      "delete": {
        "operationId": "HrProfilesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrProfileEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "HR profile not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Delete an HR profile",
        "tags": [
          "hr-profiles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wellness Platform API",
    "description": "REST APIs for the evaluation backend. Authorization is enforced using the request header role.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "Role required for RBAC. Supported values: Admin, HR, Employee, Wellness Expert"
      }
    },
    "schemas": {
...
      "EmployeeEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "employee-1777718400000-a1b2c3d4"
          },
          "name": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "email": {
            "type": "string",
            "example": "aarav.sharma@gmail.com"
          },
          "username": {
            "type": "string",
            "example": "aarav.sharma@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "secure123"
          },
          "department": {
            "type": "string",
            "example": "Engineering"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "age": {
            "type": "string",
            "example": "28"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+91 98765 43210"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "heightCm": {
            "type": "string",
            "example": "175"
          },
          "weightKg": {
            "type": "string",
            "example": "72"
          },
          "rewardPointsBalance": {
            "type": "string",
            "example": "500"
          },
          "claimedRewardIds": {
            "example": [
              "reward-1777718400000-a1b2c3d4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-02T12:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-05-02T12:45:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "username",
          "password",
          "department",
          "status",
          "age",
          "gender",
          "phoneNumber",
          "companyId",
          "companyName",
          "heightCm",
          "weightKg",
          "rewardPointsBalance",
          "claimedRewardIds",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "department": {
            "type": "string",
            "example": "Engineering"
          },
          "email": {
            "type": "string",
            "example": "aarav.sharma@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "secure123"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "age": {
            "type": "string",
            "example": "28"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "phoneNumber": {
            "type": "string",
            "example": "9876543210"
          },
          "heightCm": {
            "type": "string",
            "example": "175"
          },
          "weightKg": {
            "type": "string",
            "example": "72"
          },
          "rewardPointsBalance": {
            "type": "string",
            "example": "500"
          },
          "claimedRewardIds": {
            "example": [
              "reward-1777718400000-a1b2c3d4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "department",
          "email",
          "password",
          "companyId"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "department": {
            "type": "string",
            "example": "Engineering"
          },
          "email": {
            "type": "string",
            "example": "aarav.sharma@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "secure123"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "age": {
            "type": "string",
            "example": "28"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "phoneNumber": {
            "type": "string",
            "example": "9876543210"
          },
          "heightCm": {
            "type": "string",
            "example": "175"
          },
          "weightKg": {
            "type": "string",
            "example": "72"
          },
          "rewardPointsBalance": {
            "type": "string",
            "example": "500"
          },
          "claimedRewardIds": {
            "example": [
              "reward-1777718400000-a1b2c3d4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "HrProfileEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "hr-1777718400000-a1b2c3d4"
          },
          "name": {
            "type": "string",
            "example": "Priya Sharma"
          },
          "email": {
            "type": "string",
            "example": "priya.sharma@gmail.com"
          },
          "username": {
            "type": "string",
            "example": "priya.sharma@gmail.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+91 98765 43210"
          },
          "department": {
            "type": "string",
            "example": "Operations"
          },
          "designation": {
            "type": "string",
            "example": "HR Manager"
          },
          "location": {
            "type": "string",
            "example": "Bengaluru"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "createdBySuperUser": {
            "type": "boolean",
            "example": true
          },
          "password": {
            "type": "string",
            "example": "secure123"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-02T12:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-05-02T12:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "username",
          "phoneNumber",
          "department",
          "designation",
          "location",
          "companyId",
          "companyName",
          "status",
          "createdBySuperUser",
          "password",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateHrProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Priya Sharma"
          },
          "email": {
            "type": "string",
            "example": "priya.sharma@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "secure123"
          },
          "phoneNumber": {
            "type": "string",
            "example": "9876543210"
          },
          "department": {
            "type": "string",
            "example": "HR"
          },
          "designation": {
            "type": "string",
            "example": "HR Admin"
          },
          "location": {
            "type": "string",
            "example": "Bengaluru"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "createdBySuperUser": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UpdateHrProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Priya Sharma"
          },
          "email": {
            "type": "string",
            "example": "priya.sharma@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "secure123"
          },
          "phoneNumber": {
            "type": "string",
            "example": "9876543210"
          },
          "department": {
            "type": "string",
            "example": "HR"
          },
          "designation": {
            "type": "string",
            "example": "HR Admin"
          },
          "location": {
            "type": "string",
            "example": "Bengaluru"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "createdBySuperUser": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}